remove options builders and make options fields public
