--- sdk/cosmos/azure_data_cosmos/src/options/delete_container_options.rs
@@ -1,34 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-/// Options to be passed to [`ContainerClient::delete()`](crate::clients::ContainerClient::delete()).
-#[derive(Clone, Debug, Default)]
-pub struct DeleteContainerOptions {}
-
-impl DeleteContainerOptions {
-    /// Creates a new [`DeleteContainerOptionsBuilder`](DeleteContainerOptionsBuilder) that can be used to construct a [`DeleteContainerOptions`].
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// let options = azure_data_cosmos::DeleteContainerOptions::builder().build();
-    /// ```
-    pub fn builder() -> DeleteContainerOptionsBuilder {
-        DeleteContainerOptionsBuilder::default()
-    }
-}
-
-/// Builder used to construct a [`DeleteContainerOptions`].
-///
-/// Obtain a [`DeleteContainerOptionsBuilder`] by calling [`DeleteContainerOptions::builder()`]
-#[derive(Default)]
-pub struct DeleteContainerOptionsBuilder(DeleteContainerOptions);
-
-impl DeleteContainerOptionsBuilder {
-    /// Builds a [`DeleteContainerOptions`] from the builder.
-    ///
-    /// This does not consume the builder, and can be called multiple times.
-    pub fn build(&self) -> DeleteContainerOptions {
-        self.0.clone()
-    }
-}
